Update 03-visualization.Rmd
Changed: What reasons do you think this is? to 'What reasons can you give for this?'
Not sure that's optimal: difficult to get balance between being concise and clarity: maybe 'what reasons might explain this?' or 'can you think of reasons that might explain this?'. 
'What reasons can you think of for this?' is a bit clunky
V nice explanation of boxplots btw!