Update 05-wrangling.Rmd
In section ## Summarize variables using summarize {#summarize}: I always find it frustrating that there does not seem an easy way to include the N in the created summary, taking into account na.rm. I.e. we do want to know the N that was used for the mean and SD. Is there a way to do that? It just seems odd not to show it, as it is a key variable that influences how you interpret mean and SD, and can be useful for flagging up existence of missing data.